+++ NEWS.md
@@ -1,6 +1,7 @@
 # googleAuthR v0.7.0.9000
 
 * Change behaviour on API parse errors to write a diagnostic object to the working directory
+* Add `gar_debug_parsing` to help debug API parsing errors.
 
 # googleAuthR v0.7.0